pyproject.toml
[project]
name = "poewikibot"
version = "0.1.0"
description = "A CLI tool to query Path of Exile Wiki"
authors = [
    {name = "Nikita", email = "hekutocbi4@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "typer[all]>=0.9.0",
    "pydantic-settings>=2.2.1",
    "python-telegram-bot>=21.0",
    "httpx (>=0.28.1,<0.29.0)",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
poewiki = "poewikibot.cli:app"

[tool.poetry]
packages = [{include = "poewikibot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
